moved validations function into its own file. required it for logsController
