+++ controllers/logsController.js
@@ -2,6 +2,8 @@ const express = require("express");
 const logRoutes = express.Router();
 const logArr = require("../models/log.js");
 
+const { validateURL } = require("../models/validations.js");
+
 logRoutes.get("/", (req, res)=>{
     // *** START bonus section - Part 1
     const { order, mistakes, lastCrisis } = req.query;
@@ -61,21 +63,6 @@ logRoutes.get("/:id", (req, res)=>{
     }
 })
 
-// *** START bonus section - Part 2
-const validateURL = (req, res, next) =>{
-    // console.log("check req.body: ", typeof req.body.captainName);
-    if ((typeof req.body.captainName !== "string") ||
-        (typeof req.body.title !== "string") ||
-        (typeof req.body.post !== "string") ||
-        (typeof req.body.mistakesWereMadeToday !== "boolean") ||
-        (typeof req.body.daysSinceLastCrisis !== "number")
-    ){
-        res.json({error: "A wrong datatype was entered."});
-    }
-    next();
-};
-// *** END bonus section - Part 2
-
 // CREATE
 logRoutes.post("/", validateURL, (req, res)=>{
         // *** Part 2:
@@ -83,23 +70,5 @@ logRoutes.post("/", validateURL, (req, res)=>{
         res.json(logArr[logArr.length-1]);
     }
 )
-
-// logRoutes.post("/", (req, res)=>{
-// // *** START bonus section - Part 2
-//     // console.log("check req.body: ", typeof req.body.captainName);
-//     if ((typeof req.body.captainName !== "string") ||
-//         (typeof req.body.title !== "string") ||
-//         (typeof req.body.post !== "string") ||
-//         (typeof req.body.mistakesWereMadeToday !== "boolean") ||
-//         (typeof req.body.daysSinceLastCrisis !== "number")
-//     ){
-//         res.json({error: "A wrong datatype was entered."});
-//     } else {
-//         // *** Part 2:
-//         logArr.push(req.body);
-//         res.json(logArr[logArr.length-1]);
-//     }
-//     // *** END bonus section - Part 2
-// })
 
 module.exports = logRoutes;
